scrapeGithub/requirements.txt
antiorm==1.2.1
attrs==21.2.0
Automat==20.2.0
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
cssselect2==0.4.1
db==0.1.1
db-sqlite3==0.0.1
decorator==5.0.9
defusedxml==0.7.1
ghost==0.6.1
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
itemadapter==0.3.0
itemloaders==1.0.4
jmespath==0.10.0
jsonschema==3.2.0
lxml==4.6.3
numpy==1.21.1
parsel==1.6.0
Pillow==8.3.1
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
pyrsistent==0.18.0
queuelib==1.6.1
requests==2.26.0
Scrapy==2.5.0
selenium==3.141.0
service-identity==21.1.0
six==1.16.0
tinycss2==1.1.0
tinydb==4.5.1
Twisted==21.7.0
typing-extensions==3.10.0.0
urllib3==1.26.6
validators==0.18.2
w3lib==1.22.0
webencodings==0.5.1
zope.interface==5.4.0
